just a start on the bulk insert/update/delete interface
The bulk() method should return objects for all of the documents that
are inserted.  The document objects that were passed as "delete" and
"update" arguments should be modified in place so that they represent
the newly changed documents.

Perhaps call a method like

    $doc->_you_are_now({ rev => $rev, deleted => 1 })

to tell the documents about their new state.  Document objects should
also have a new method is_deleted() to indicate whether or not they
represent a deleted object.  Try running t/db-bulk.t with DEBUG=1 to
see the HTTP response body and how to proceed.

there should definitely be sugary

    my @new_docs = $db->insert(@hashes);
    $db->delete(@documents)
    $db->update(@documents)

That reminds me, I need to make sure that doing a bunch of operations
in a single bulk() invocation doesn't accomplish anything if there is
one error.