A new way for making requests to the CouchDB server
The new method request() consolidates much of the code that
is necessary every time we make a request to the server.
This should simplify most methods making it clearer what the
method is actually trying to accomplish.  It will also
make error messages more consistent.

The old call() interface is officially deprecated and generates
ugly warnings if it's used.

Signed-off-by: Michael Hendricks <michael@ndrix.org>