Make request() a mixin method
Many classes will want to use this logic for making requests to
the CouchDB instance.  Using inheritance to make the method
available is wrong because, for instance, a Document is not a DB.

Ideally, Net::CouchDB::Request would be implemented with Class::Trait
or Moose::Role so that the method requirements would be handled
automatically.  I just can't stomach adding a dependency for that
when the user's of this module don't really get anything for it.

Signed-off-by: Michael Hendricks <michael@ndrix.org>