Support the new _bulk_docs API
As of r753448, CouchDB has changed the semantics and interface for the
_bulk_docs API.  It no longer does much conflict checking.  Successful
document updates are now reported as an arrayref of hashes.  It used
to be reported as a hash in which the new_revs key contained an
arrayref of hashes.

This makes all the tests pass again for the current CouchDB trunk.

Signed-off-by: Michael Hendricks <michael@ndrix.org>